+++ api/plugin-api/src/main/java/com/yanncebron/m68kplugin/browser/M68kBrowserPaneFactoryEP.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 The Authors
+ * Copyright 2026 The Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,7 +24,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.ResourceBundle;
 
-public final class M68kBrowserPaneFactoryEP extends BaseKeyedLazyInstance<M68kBrowserPaneFactory<?>> {
+public final class M68kBrowserPaneFactoryEP extends BaseKeyedLazyInstance<M68kBrowserPaneFactory<?, ?>> {
 
   @RequiredElement
   @Attribute("displayNameKey")